TrashCan Object
Added a second trash can to the inner of the trash can object. Currently it is two mesh colliders within each other. The inner most one is a trigger that will deactivate the object and count the points when I make the script (has the tag "caught"). The other is a non-trigger collider so the items can bounce off the can.